+++ controllers/index.js
@@ -1,4 +1,5 @@
 module.exports = {
 	homeController: require("./home"),
 	userController: require("./user"),
+	privateController: require("./private"),
 };
