+++ routers/index.js
@@ -4,5 +4,6 @@ const router = express.Router();
 router
     .use("/", require("./home"))
     .use("/user", require("./user"))
+    .use("/private", require("./private"))
 
 module.exports = router;
