+++ templates/1_Public/11_who.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is who we are.</h1>
+
+
+${footer}
