+++ templates/1_Public/12_where.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is where we work.</h1>
+
+
+${footer}
