+++ templates/1_Public/13_blog.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is our blog.</h1>
+
+
+${footer}
