+++ templates/1_Public/14_cookies.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is our cookie policy.</h1>
+
+
+${footer}
