+++ templates/1_Public/15_privacy.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is our privacy policy.</h1>
+
+
+${footer}
