+++ templates/1_Public/16_contact.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is how you can contact us.</h1>
+
+
+${footer}
