+++ templates/1_Public/17_careers.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! This is how you can join us.</h1>
+
+
+${footer}
