+++ templates/2_Private/21_chat-rules.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! These are our ANGO-Chat rules.</h1>
+
+
+${footer}
