+++ templates/2_Private/21_chat.html
@@ -1,0 +1,7 @@
+${header}
+
+<h1>Welcome ${username} to ANGO-Chat.</h1>
+
+<a href="/members/create"><button>ADD POST</button></a>
+
+${footer}
