+++ templates/2_Private/22_abod.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here's a message from our ANGO Board of Governors.</h1>
+
+
+${footer}
