+++ templates/2_Private/22_aoff.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here's a message from our ANGO Officers.</h1>
+
+
+${footer}
