+++ templates/2_Private/23_cbod.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here's a message from our ${country} Board of Directors.</h1>
+
+
+${footer}
