+++ templates/2_Private/23_coff.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here's a message from our ${country} Officers.</h1>
+
+
+${footer}
