+++ templates/2_Private/24_vote.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here is information about our election process.</h1>
+
+
+${footer}
