+++ templates/2_Private/25_conferences.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here is information about our conferences.</h1>
+
+
+${footer}
