+++ templates/2_Private/26_calendar.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here is our calendar.</h1>
+
+
+${footer}
