+++ templates/2_Private/27_travel.html
@@ -1,0 +1,6 @@
+${header}
+
+<h1>Welcome ${username}! Here are travel tips.</h1>
+
+
+${footer}
