+++ templates/deleteComment.html
@@ -1,11 +1,8 @@
 ${header}
-<ul class="topnav">
-    <li><a href="/private">Home</a></li>
-    <li><a href="/logout">Logout</a></li>
-  </ul>
+
 <form method="POST">
     <h2>
-        Would you like to delete this comment? ${comment.id}
+        Would you like to delete this comment?
         
     </h2>
 <br>
