+++ templates/deletePost.html
@@ -1,11 +1,8 @@
 ${header}
-<ul class="navbar">
-    <li><a href="/private">Home</a></li>
-    <li><a href="/logout">Logout</a></li>
-  </ul>
+
 <h1>${name}</h1>
 <form method ="POST">
-    <h2>Would you like to delete this Post: ${post.id}?</h2>
+    <h2>Would you like to delete this Post?</h2>
     <br>
     <input type="submit" value="submit">
 </form>
