+++ templates/editComment.html
@@ -1,8 +1,5 @@
 ${header}
-<ul class="navbar">
-    <li><a href="/private">Home</a></li>
-    <li><a href="/logout">Logout</a></li>
-  </ul>
+
 <h1>Comment on: ${post.title}</h1>
 <div>
     ${post.media.includes("/uploads/media/") ? `<img src="${post.media}" alt="${user.username}'s image" />`:'' }
