+++ templates/editPost.html
@@ -1,8 +1,4 @@
 ${header}
-<ul class="navbar">
-    <li><a href="/private">Home</a></li>
-    <li><a href="/logout">Logout</a></li>
-</ul>
 
 <form method="POST" action="/post/${post.id}/edit" enctype="multipart/form-data">
     <label >Title 
@@ -17,7 +13,7 @@ ${header}
         <textarea  rows="4" cols="50" name="content" >${post.content}</textarea>
     </label> 
     <br>
-    <input type="submit" value="POST" >
+    <input type="submit" value="submit" >
 </form>
 <script>
     console.log(post.id)
