+++ templates/home.html
@@ -1,4 +1,4 @@
-${header}
+<!-- ${header}
 <div class="container">
 <ul class="navbar">
     <li><a href="/">Home</a></li>
@@ -9,4 +9,126 @@ ${header}
 <br>
 <a href="/user/signup">Signup</a>     or     <a href="/user/login">Login</a>
 </div>
+${footer} -->
+
+${header}
+<div class="container">
+    <div class="row">
+        <div class="col-sm-2">
+            <h3><a href="/"></a></h3>
+            <a href="/user/signup">Signup</a> or <a href="/user/login">Login</a>
+        </div>
+        <div class="col-sm-8">
+            <h6><a href="/">Cookie policy - This site uses cookies, and by continuing to browse this site we assume you agree to this. You can turn cookies off within your browser, although certain parts of the site will not work if you do. For more information please read our cookie policy.</a></h6>
+            <p></p>
+        </div>
+        <div class="col-sm-2">
+            <h6></h6>
+            <button type="button" onclick="alert('Cookies')">Accept</button>
+        </div>
+    </div>
+</div>
+<div class="container">
+    <div class="row">
+        <div class="col-sm-3">
+            <h3><a href="/">Language</a></h3>
+            <p>[Dropdown Menu]</p>
+        </div>
+        <div class="col-sm-6">
+            <h3><a href="/">Asian Non Governmental Organization</a></h3>
+            <p>(ANGO)</p>
+        </div>
+        <div class="col-sm-3">
+            <h3><a href="/">Donate Now</a></h3>
+            <p>[Link to donation page]</p>
+        </div>
+    </div>
+</div>
+<div class="container">
+    <div class="row">
+        <div class="col-sm-2">
+            <h5><a href="/">Search</a></h5>
+            <p>[Open Search Box]</p>
+        </div>
+        <div class="col-sm-3">
+            <h3><a href="/">Who we are</a></h3>
+            <ul class="col-sm-8">
+                <li><a href="/Vision">Vision</a></li>
+                <li><a href="/Mission">Mission</a></li>
+                <li><a href="/History">History</a></li>
+                <li><a href="/Board_Members">Board Members</a></li>
+                <li><a href="/Field_of_Work">Field of Work</a></li>
+                <li><a href="/Governance">Governance</a></li>
+                <li><a href="/Standards">Standards</a></li>
+                <li><a href="/Finances">Finances</a></li>
+            </ul>
+        </div>
+        <div class="col-sm-3">
+            <h3><a href="/">Where we work</a></h3>
+            <p>[Link to countries page]</p>
+        </div>
+        <div class="col-sm-2">
+            <h3><a href="/">Blog</a></h3>
+            <ul class="col-sm-4">
+                <li><a href="/News">News</a></li>
+                <li><a href="/Publications">Publications</a></li>
+            </ul>
+        </div>
+        <div class="col-sm-2">
+            <h5><a href="/">More</a></h5>
+            <ul class="col-sm-4">
+                <li><a href="/Careers">Careers</a></li>
+                <li><a href="/Contact">Contact</a></li>
+                <li><a href="/Privacy">Privacy</a></li>
+                <li><a href="/Cookies">Cookies</a></li>
+            </ul>
+        </div>
+    </div>
+</div>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<br>
+<div class="container">
+    <div class="row">
+        <div class="col-sm-3">
+            <h3><a href="/"></a></h3>
+                <p><a href="/Vision">Vision</a></p>
+                <p><a href="/Mission">Mission</a></p>
+                <p><a href="/History">History</a></p>
+                <p><a href="/Board_Members">Board Members</a></p>
+        </div>
+        <div class="col-sm-3">
+            <h3><a href="/"></a></h3>
+                <p><a href="/Field_of_Work">Field of Work</a></p>
+                <p><a href="/Governance">Governance</a></p>
+                <p><a href="/Standards">Standards</a></p>
+                <p><a href="/Finances">Finances</a></p>
+        </div>
+        <div class="col-sm-3">
+            <h3><a href="/"></a></h3>
+                <p><a href="/News">News</a></p>
+                <p><a href="/Publications">Publications</a></p>
+        </div>
+        <div class="col-sm-3">
+            <h3><a href="/"></a></h3>
+                <p><a href="/Careers">Careers</a></p>
+                <p><a href="/Contact">Contact</a></p>
+                <p><a href="/Privacy">Privacy</a></p>
+                <p><a href="/Cookies">Cookies</a></p>
+        </div>
+    </div>
+</div>
+<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
 ${footer}
