+++ templates/login.html
@@ -1,9 +1,4 @@
 ${header}
-<ul class="navbar">
-  <li><a href="/">Home</a></li>
-  <li><a href="/about">About</a></li>
-  <li><a href="/contact">Contact</a></li>
-</ul>
 <h1>${title}</h1>
 <form action="" method="post">
     <label for="username">Username:</label>
