+++ templates/signup.html
@@ -1,8 +1,4 @@
 ${header}
-<ul class="navbar">
-  <li><a href="/">Home</a></li>
-  <li><a href="/about">About</a></li>
-  <li><a href="/contact">Contact</a></li>
 <h1>${title}</h1>
 <div class="container">
 <form action="" method="post">
